requirements.txt
configparser==3.5.0
cycler==0.10.0
future==0.16.0
jedi==0.11.1
json-rpc==1.10.8
matplotlib==2.1.2
mccabe==0.6.1
numpy==1.14.0
oandapyV20==0.4.5
pandas==0.22.0
parso==0.1.1
pkg-resources==0.0.0
pluggy==0.6.0
pyaml==17.12.1
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-language-server==0.13.0
pytz==2017.3
PyYAML==3.12
rope==0.10.7
six==1.11.0
snowballstemmer==1.2.1
yapf==0.20.1
